C1 - Initial commit
This commit provides a base CRUD for the core service entities.

Fun fact: this initial commit is part of a 24-hour session made to discuss, plan and initiate the project. I wish I had written more code but hey, for a code made under the influence of energy drinks, I guess it wasn't all that bad.