C2 - Use Docker to deploy the project
Used Docker so the Frontend team can deploy and test the project.
The core service resides under its own container.
To test the database locally, a container with a PostgresSQL database was added.
Although it's not currently necessary, since it's planned to create a microservices architecture, an nginx container was created to route traffic to the appropiate service.