C4 - Added pagination to hostels
Modified the GET /hostel endpoint so it can receive query parameters, incluiding one that allows filtering according to the services that the hostel provides. It also has query parameters to search between a time period while also returning the available number of spaces for each hostel.
Fun fact: This is the first time I do filtering on a many-to-many relation, the query looks quite scary, thanks to chatgpt for existing! Let's just hope the examiner doesn't read this commit comment though...