C5 - Add Persons logic and minor changes
To store information regarding each person's health included in a reservation, the user can now add profiles of people which will be saved in the "persons" table. To relate a person to a reservation, an auxiliary table "person_reservations" was created. Since the table "person_reservations" can be used to determine the number of people in a reservation, the column "people_count" was eliminated from the "reservations" table. The hostel pagination query was also modified due to this change.
A minor change in the base repository template was made and how many services used a repository.
A bug for getting the imageUrls for a hostel in which an incorrect variable was used was fixed.
Fun fact: I'm starting to worry because I feel I write code too slowly, I don't know if it's because it's the first time I use Spring Boot or because Java has waaaaay too much boilerplate code...