C7 - Added necessary configuration for AWS
Divided configuration files between dev, staging and prod configurations which was necessary for uploading the project to AWS. This configuration consist in the (added in a previous C2) Dockerfiles, docker-compose, and nginx.conf files.
Removed the unnecessary named volumes in the core service for staging and development environments.
Fun fact: Configuring nginx and HTTPS was surely fun, specially when we take into account Docker...
I wonder how much funner it will get when i get to the implementation of the CD pipeline.