C8 - Added the reservation history API
Added the endpoints for getting the reservation history and repeating a reservation.
Modified the response for the creation of reservations so it now includes the IDs of the people and services included in the reservation.
Fun fact: Today we had a session of feedback with the client, seems like we are going to have lots of fun changing some of the requierements, between this and the fact that this is my first time in Java and Spring Boot, I think I'll be spending most of my time refactoring code. Well, at least I'm learning...