C11 - Make changes to the user's schema
Since the user will now require a phone number instead of an email, the email field was removed to reduce the amount of data asked from the user.
Modified the endpoint that gets a user's reservations to obtain the user's ID from a header so it can integrate with API Gateway.
Made the pertinent changes in the script to fill the database to accomodate the removal of the email field.
Added a folder to the gitignore file named "local" to put commodity scripts that depend on things like the OS of the machine.
Fun fact: This is a small commit because the most of the changes were on the AWS part, but overall, there were less changes than I expected, let's hope testing goes as quick.