C12 - Service reservation system and other stuff
Added the endpoints necessary to create pending service reservations and how to confirm them.
Created a table for each service that hostels provide.
Created the dispatcher that calls the API to create external service reservations.
Added endpoint that allows to get the currently active reservation of an user.
Remove dependency on the core container for the nginx conatiner.
Fun fact: This commit includes a lot of boilerplate code. I've been considering seriously to start using a strong IA tool such as Cursor or Claude Code but I have my doubts, I know I have to eventually learn to use these tools, but I'm a bit worried about losing control of the code I ship. Perhaps on the next project.