C13 - Change userId endpoints and more
Endpoints that require a user ID in the mobile app send this ID through a header, this was done previously in the user reservation endpoint, this needed to be replicated to other endpoints.
The price of the hostel was added to the hostels table instead of being stored as a value in the services table.
In the hostel pagination endpoint, only the first hostel image is sent.
For testing purposes, the default state for a reservation was the active state, that was changed to the pending state.
Made the necessary changes to the fillDB.py script and added random generation of previous reservations and also for service reservations.
Added a local login endpoint for testing which needed a new method in the userRepository.
Fun fact: We are running out of time and I still need to create some endpoints for the admin web page, I hope this is the last major mobile commit...