C14 - Add dashboard endpoints
Added the endpoints for the dashboard to track number of reservations, number of people, state of reservations and types of services.
Modified the fillDB.py script to increase the number of users created to 8 and for the possibility to create a pending reservation.
Fun fact: By this point, all major endpoints of the backend should done, let's hope any future commits are minor or optional changes, such as the Continuous Deployment pipeline I've been wanting to do.