C15 - Separate regular and admin routes and more
Separated regular and admin routes in two different controllers.
Changed the route prefix of the services endpoints.
Made the x-user-id header optional so it worked on development and production.
Added exception handling so error messages are more understandable.
Fixed bug that allowed a user to have many pending reservations.
Fixed bug where it would be mandatory to send all data of an object to update it.
Added validations to the fields of the endpoints' requests.
Changed the routes of some endpoints such as the reservations' dashboard, get a user's active reservation and get a user's reservation history.
Added a local signup and signup confirmation to do better local tests.
Removed an unnecessary env variable "DB_NAME" and set a fixed name for the DB.
Added environment variable to the README.md setup.
Fun fact: Ok I lied, I had some more changes to do, but they are mostly minor changes, it's just that I put them most of them in a single commit. I still had some more changes to do but they  will be fewer than this commit.